backend/requirements.txt
aggdraw==1.3.11
asgiref==3.2.10
attrs==19.3.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
Django==3.0.8
django-cors-headers==3.4.0
django-extensions==3.0.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
imageio==2.9.0
kiwisolver==1.2.0
matplotlib==3.3.0
networkx==2.4
numpy==1.19.0
packbits==0.6
Pillow==7.2.0
psd-tools==1.9.15
PyJWT==1.7.1
pymaging-psd==0.7
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
qrcode==6.1
scikit-image==0.16.2
scipy==1.5.1
six==1.15.0
sqlparse==0.3.1
